+++ changelog.md
@@ -1,3 +1,9 @@
+# v8.0.0-beta.26
+- Update README.md
+- Warn for zip related functions on android
+- Force showing for LevelSettingsLayer (android bug)
+- Fix listing setting update when you add level. (insert tool)
+
 # v8.0.0-beta.25
 fix recursion at level save